circuit: Rename transaction -> tx_circuit
